pyproject.toml
[tool]
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Marco Acierno", email = "marcoaciernoemail@gmail.com"},
]
dependencies = [
    "django>=4.1.7",
    "strawberry-graphql>=0.158.1",
]
requires-python = ">=3.11"
license = {text = "MIT"}
